# src/API/AhcmsAPI/Models/Place.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;

namespace AhcmsAPI.Models
{
    public class Place
    {
        public int Id { get; set; }
        public string Location { get; set; }
    }
}